requirements.txt
flask==1.1.2
Pillow
Keras
numpy
tensorflow>=2.2
https://github.com/lhelontra/tensorflow-on-arm/releases/download/v2.2.0/tensorflow-2.2.0-cp37-none-linux_armv7l.whl ; platform_machine == 'armv7l'
